example-lambda-api/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-lambda-powertools>=2.25.0
pydantic>=2.0.0
pytest>=7.0.0